Rename esperanza/init.* files
